Factors out colors to utils dir
