+++ client/modules/reconportal/src/LeafletMap/leafletUtil.tsx
@@ -2,8 +2,6 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import {
-  OpenTopoDataset,
-  ReconPortalEvents,
   useReconEventContext,
   getReconPortalEventIdentifier,
 } from '@client/hooks';
@@ -55,47 +53,6 @@ export const createClusterIcon = (cluster: MarkerCluster) => {
     ),
   });
 };
-
-/**
- * Get color for OpenTopo dataset
- */
-export function getOpenTopoColor(dataset: OpenTopoDataset): string {
-  return 'black';
-}
-
-/**
- * Declare and get color for Recon Portal dataset
- */
-export const EVENT_TYPE_COLORS = {
-  earthquake: ' #D34141',
-  flood: ' #5D1393',
-  tsunami: ' #CCC526',
-  landslide: ' #57D657',
-  hurricane: ' #47A59D',
-  tornado: ' #FF5722',
-};
-
-export function getReconEventColor(
-  dataset: ReconPortalEvents | undefined | null
-): string {
-  if (!dataset) return '#666666'; // Default gray color
-  switch (dataset?.event_type) {
-    case 'earthquake':
-      return EVENT_TYPE_COLORS.earthquake;
-    case 'flood':
-      return EVENT_TYPE_COLORS.flood;
-    case 'tsunami':
-      return EVENT_TYPE_COLORS.tsunami;
-    case 'landslide':
-      return EVENT_TYPE_COLORS.landslide;
-    case 'hurricane':
-      return EVENT_TYPE_COLORS.hurricane;
-    case 'tornado':
-      return EVENT_TYPE_COLORS.tornado;
-    default:
-      return '#666666'; // Default gray color
-  }
-}
 
 /**
  * A React-Leaflet wrapper that conditionally renders its children (e.g. markers, GeoJSON, etc.)
