+++ client/modules/reconportal/src/index.ts
@@ -1,2 +1,3 @@
 export * from './ReconPortal/ReconPortal';
 export * from './ReconSidePanel/ReconSidePanel';
+export * from './utils/colors';
